using BATMASS library for parsing mzML file.

+++ mspp/src/main/java/ninja/mspp/model/entity/PeakList.java
@@ -1,13 +1,20 @@
 package ninja.mspp.model.entity;
 
 import java.io.Serializable;
-import javax.persistence.*;
 import java.util.List;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.NamedQuery;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
 
 
 /**
  * The persistent class for the PEAK_LISTS database table.
- * 
+ *
  */
 @Entity
 @Table(name="PEAK_LISTS")
@@ -17,7 +24,7 @@ public class PeakList implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 
 	//bi-directional many-to-one association to Peak
 	@OneToMany(mappedBy="peakList")
@@ -26,11 +33,11 @@ public class PeakList implements Serializable {
 	public PeakList() {
 	}
 
-	public int getId() {
+	public Long getId() {
 		return this.id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
